requirements.txt
Django==3.2.*
pytz
Jinja2==2.11.*
Pillow==8.4.0
